event-auditor: code generation
It possible to describe a group of ipv4 addresses using globs,
ranges and CIDR notation. Internally, we convert the CIDR notation
to range notation before tokenize the value.

A list of port numbers can be described using numbers or ranges.

The mode parameter can be described as a number of a sequence of
bitwise ORed numbers (using pipe). E.g.: 0x400 | 0x02 | 0x10.

For the sake of simplicity, I avoided to create a fully parsing
mechanism. The current code tokenize all parameters and apply a direct
translation on these (based on a list of formalized rules).

The main limitation of this approach is that we lost any kind of
expression analysis. Meaningless match constructions like:

ipv4addr: 192.168.*.*, -192.168.*.*

will go unnoticed, simply because the code is not able to understand
the expression as a whole.