event-auditor: use uniq program name to avoid conflicts
When some rule is updated in a way that affects the ebpf program, a new
program is generated, compiled and loaded. The old program however is
not immediately removed (it still cached and is removed by some garbage
collection mechanism - not yet implemented).

The previous logic has tried to use the same program names for both
the old and new programs. This commit fix this problem ensuring a uniq
program name on all generated code.