Add bpf tracepoint for sys_exit_openat event
This commit adds bpf tracepoint sys_exit_openat to compensate for untriggered kretprobes for openat syscall. This tracepoint will act as a double check.
Compared to other syscalls openat is the most frequent and the kretprobe count is much less than entry kprobe which resulted in a lot of untraced events.
This significantly reduces missing logs and event count in KubeArmor.

Fixes: #442

Signed-off-by: Ayush Dwivedi <ayush.dwivedi@accuknox.com>