apparmor-enforcer: skip non-regular profiles
SUSE creates few symbolic links in the `/etc/apparmor.d/` folder.
Kubearmor gives out error since the current check if based on IsDir()
and the sym links are not covered in that check. Now we are checking if
the file is a regular file and only use it in that case.

Signed-off-by: Rahul Jadhav <nyrahul@gmail.com>