fixes k8s-nodeName, hostname mismatch issue
Kubearmor goes into CrashLookBackOff if for some reason the k8s-nodeName
does not match with the node's hostname.
k8s-nodename is usually derived from the node's hostname, but if for
some reason the node's hostname is changed then the k8s-nodename is not
correspondingly changed. This causes the kubearmor to go in a crash
loop while printing the message "The node information is not available".

**Why does Kubearmor need to match nodename and hostname?**
Kubearmor needs to set annotation on the local node. For that it needs
to know the node's metadata from k8s and match it up with the hostname.

Signed-off-by: Rahul Jadhav <nyrahul@gmail.com>