gracefully terminate kubearmor container
When k8s or CRI stop containers, they send SIGTERM to pid 1 in the container. pid 1 in this case is bash which executes our script and never really passed the signal onto child processes for it to gracefully handle and waited for the termination timeout to get abruptly killed and hence not cleaning up resources.

We now leverage the binary directly as entrypoint to put KubeArmor process as PID 1 so KubeArmor directly handles the termination signals.

Now KubeArmor will clean up resources while being terminated, also the process will be quicker since we don't wait for the entire period for it to get abruptly killed later.

Signed-off-by: daemon1024 <barun.acharya@accuknox.com>