core: update visibility setting from configmap
This commit updates the way we handle visibility annotations.

We now use configmap/global config value as the default source of truth and don't annotate each and every namespace.

Any changes to the configmap will update  all the namespaces visibility that doesn't have annotation.

If anyone needs to set a custom visibility annotation for some namespace they can annotate the namespace and it will overrule the global default

Signed-off-by: daemon1024 <barun1024@gmail.com>