fix(CI): add the job to run go test on the KubeArmor/KubeArmor directory and fix to pass the tests
The current implementation does not have a job in CI to run tests for
golang programs under the KubeArmor/KubeArmor directory.

Therefore, this commit will add the job to ci-test-go.yml to run tests for
golang programs under the KubeArmor/KubeArmor directory.

On the other hand, in the current implementation, the following error
occurs when go test is executed in the directory KubeArmor/KubeArmor/core.

  > go test
  # github.com/kubearmor/KubeArmor/KubeArmor/core
  ./karmorprobedata.go:61:3: (*github.com/kubearmor/KubeArmor/KubeArmor/feeder.Feeder).Errf format %s reads arg #1, but call has 0 args

Thus, if test_kubearmor.sh or go test ./... on the kubearmor/kubeArmor are
run, the test will fail.

The test fails because dm.Logger.Errf(), which is called in the
SetKarmorData function in the file core/karmorprobedata.go, is missing an
argument for the format specifier.

Therefore, this commit will fix it to add a variable for the format
specifiers so that the test will succeed properly.

Signed-off-by: Hayato Kiwata <simplelpmis6@gmail.com>