test(syscalls): Add an annotation to Pod for syscalls test
When we run ginkgo -r in the KubeArmor/tests/syscalls directory, we get a
Pod not found error even though the Pod for testing is running, as shown
below. Note that the error message is excerpted.

------------------------------
• [FAILED] [61.257 seconds]
Syscalls [BeforeEach] Match syscalls can detect unlink syscall
  [BeforeEach] /home/ec2-user/KubeArmor/tests/syscalls/syscalls_test.go:49
  [It] /home/ec2-user/KubeArmor/tests/syscalls/syscalls_test.go:59

  [FAILED] Expected
      <*errors.errorString | 0xc0006dc1e0>:
      pod not found
      {s: "pod not found"}
  to be nil
  In [BeforeEach] at: /home/ec2-user/KubeArmor/tests/syscalls/syscalls_test.go:40 @ 07/02/23 06:31:59.513
------------------------------

This is an error in K8sGetPods() in syscalls_test.go because the Pod
deployed by ubuntu-deployment.yaml in the KubeArmor/tests/syscalls/manifests
directory was not found. The current implementation of the test is not
working.

The current implementation of the test uses the pod name prefix and
namespace and annotation to find the Pod that will execute the test commands.
In the implementation in syscalls_test.go, the prefix, namespace and
annotation are "ubuntu-1-deployment-", "syscalls", "kubearmor-policy: enabled".

The Pod being searched for is not annotated, and the test will not run
because it cannot find the Pod even though it is running, resulting in
the error described above.

Therefore, this commit modifies ubuntu-deployment.yaml to annotate the Pod
so that the test will work correctly.

Signed-off-by: Hayato Kiwata <simplelpmis6@gmail.com>