fix(core): fix to prevent policies from being applied to containers that shouldn't be applied
When running KubeArmor in systemd mode, when the container name (the value
of kubearmor.io/container.name) in the applied policy was changed and the
policy was applied, there was an issue where that policy was applied not
only to the new container but also to the previous container.

This bug was also reported in issue/#1387.

Therefore, in this commit, the policy doesn't apply to the previous
container, and apply to the new container.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>