fix(core): fix so that it can be restricted to matched paths and directories in systemd mode
When running KubeArmor in systemd mode, there is an issue where
restrictions on the directories specified in sepc.file.matchDirectories
are not enforced when matchPaths and matchDirectories are applied to
spec.file in HostPolicy as shown below.

  ```
  apiVersion: security.kubearmor.com/v1
  kind: KubeArmorHostPolicy
  metadata:
    name: host-policy
  spec:
    file:
      matchPaths:
      - path: /home/ubuntu/.bashrc
      matchDirectories:
      - dir: /etc/cron.weekly/
        recursive: true
    action:
      Block
  ```

This issue is reported in issue/#1436.

In the current implementation, it is possible to check files under
"/etc/cron.weekly/" directory even if this Policy is applied.

Therefore, in this fix, the implementation was made to fix this issue.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>