feat(monitor): monitor the chroot systemcall
Currently, KubeArmor does not have the feature to monitor the chroot
system call.

On the other hand, there is a future request in Issue #1070 to support
audit for security sensitive system calls.

  - Issue #1070 : support audit for security sensitive system calls
    - #1070

Therefore, this commit implements a new feature to monitor the chroot
system call.

This commit also includes a corresponding fix to allow standard input to
be used in the K8sExecInPod function in kartutil.go to run ginkgo test for
the chroot system call and so on.

Signed-off-by: Hayato Kiwata <haytok@amazon.co.jp>