feat(bpf): use LRU_HASH instead of HASH maps
We keep certain maps to track processes and containers and there respective args/file data on various syscalls. These could expand a lot and could result in filling up of the maps due to edge case failure or potential not triggering of the kretprobe.

By using an LRU_HASH we ascertain that even if the map is filled up it will starts evicting least recently used ones, so as to make sure current monitoring is not affected

Signed-off-by: daemon1024 <barun1024@gmail.com>