refactor(monitor/bpf): performance enhancement
We used to drop system events only at kretprobe earlier because we needed to confirm that the event is not an alert before dropping it. But alerts are handled by the bpflsm enforcer on it's own.
This commit adds a check if the enforcer is BPFLSM we can drop events as soon as the hook is triggered instead of waiting for the entire context in kretprobe.

We also updated the default case of drop syscall logic. Earlier the default case of tracing syscalls if we can't find the respective container visibility map or details.
But there's a lag between when we start receiving events for a particular container and the creation of the container map because of latency in event from CRI socket and other userspace latencies.
This caused us to send file events to userspace even if file events are disabled in general.
We now try to check global config to check if lookup for container fails do we need to ignore or trace a syscall and In case lookup for global config fails we continue the container check and choose to not drop events if either of them are not found.

Signed-off-by: daemon1024 <barun1024@gmail.com>